Move hashing_helper to utils/ dir.
