+++ gslib/copy_helper.py
@@ -69,15 +69,6 @@ from gslib.daisy_chain_wrapper import DaisyChainWrapper
 from gslib.exception import CommandException
 from gslib.exception import HashMismatchException
 from gslib.file_part import FilePart
-from gslib.hashing_helper import Base64EncodeHash
-from gslib.hashing_helper import CalculateB64EncodedMd5FromContents
-from gslib.hashing_helper import CalculateHashesFromContents
-from gslib.hashing_helper import CHECK_HASH_IF_FAST_ELSE_FAIL
-from gslib.hashing_helper import CHECK_HASH_NEVER
-from gslib.hashing_helper import ConcatCrc32c
-from gslib.hashing_helper import GetDownloadHashAlgs
-from gslib.hashing_helper import GetUploadHashAlgs
-from gslib.hashing_helper import HashingFileUploadWrapper
 from gslib.parallel_tracker_file import GenerateComponentObjectPrefix
 from gslib.parallel_tracker_file import ReadParallelUploadTrackerFile
 from gslib.parallel_tracker_file import ValidateParallelCompositeTrackerData
@@ -128,6 +119,15 @@ from gslib.utils.encryption_helper import CryptoKeyType
 from gslib.utils.encryption_helper import CryptoKeyWrapperFromKey
 from gslib.utils.encryption_helper import FindMatchingCSEKInBotoConfig
 from gslib.utils.encryption_helper import GetEncryptionKeyWrapper
+from gslib.utils.hashing_helper import Base64EncodeHash
+from gslib.utils.hashing_helper import CalculateB64EncodedMd5FromContents
+from gslib.utils.hashing_helper import CalculateHashesFromContents
+from gslib.utils.hashing_helper import CHECK_HASH_IF_FAST_ELSE_FAIL
+from gslib.utils.hashing_helper import CHECK_HASH_NEVER
+from gslib.utils.hashing_helper import ConcatCrc32c
+from gslib.utils.hashing_helper import GetDownloadHashAlgs
+from gslib.utils.hashing_helper import GetUploadHashAlgs
+from gslib.utils.hashing_helper import HashingFileUploadWrapper
 from gslib.utils.metadata_util import ObjectIsGzipEncoded
 from gslib.utils.posix_util import ATIME_ATTR
 from gslib.utils.posix_util import ConvertDatetimeToPOSIX
