+++ gslib/tests/test_copy_helper_funcs.py
@@ -19,6 +19,7 @@ import logging
 import os
 
 from apitools.base.py import exceptions as apitools_exceptions
+import mock
 
 from gslib.cloud_api import ResumableUploadAbortException
 from gslib.cloud_api import ResumableUploadException
@@ -35,7 +36,6 @@ from gslib.copy_helper import PerformParallelUploadFileToObjectArgs
 from gslib.copy_helper import WarnIfMvEarlyDeletionChargeApplies
 from gslib.discard_messages_queue import DiscardMessagesQueue
 from gslib.gcs_json_api import GcsJsonApi
-from gslib.hashing_helper import CalculateB64EncodedMd5FromContents
 from gslib.parallel_tracker_file import ObjectFromTracker
 from gslib.storage_url import StorageUrlFromString
 from gslib.tests.mock_cloud_api import MockCloudApi
@@ -47,8 +47,10 @@ from gslib.third_party.storage_apitools import storage_v1_messages as apitools_m
 from gslib.utils import parallelism_framework_util
 from gslib.utils import posix_util
 from gslib.utils import system_util
+from gslib.utils import hashing_helper
 
-import mock
+_CalculateB64EncodedMd5FromContents = (
+    hashing_helper.CalculateB64EncodedMd5FromContents)
 
 
 class TestCpFuncs(GsUtilUnitTestCase):
@@ -114,7 +116,7 @@ class TestCpFuncs(GsUtilUnitTestCase):
         self.default_provider, bucket_name,
         fpath_uploaded_correctly))
     with open(fpath_uploaded_correctly) as f_in:
-      fpath_uploaded_correctly_md5 = CalculateB64EncodedMd5FromContents(f_in)
+      fpath_uploaded_correctly_md5 = _CalculateB64EncodedMd5FromContents(f_in)
     mock_api.MockCreateObjectWithMetadata(
         apitools_messages.Object(bucket=bucket_name,
                                  name=fpath_uploaded_correctly,
@@ -144,7 +146,7 @@ class TestCpFuncs(GsUtilUnitTestCase):
     object_wrong_contents_url = StorageUrlFromString('%s://%s/%s' % (
         self.default_provider, bucket_name, fpath_wrong_contents))
     with open(self.CreateTempFile(contents='_')) as f_in:
-      fpath_wrong_contents_md5 = CalculateB64EncodedMd5FromContents(f_in)
+      fpath_wrong_contents_md5 = _CalculateB64EncodedMd5FromContents(f_in)
     mock_api.MockCreateObjectWithMetadata(
         apitools_messages.Object(bucket=bucket_name,
                                  name=fpath_wrong_contents,
@@ -167,7 +169,7 @@ class TestCpFuncs(GsUtilUnitTestCase):
     # Exists in tracker file and already uploaded, but no longer needed.
     fpath_no_longer_used = self.CreateTempFile(file_name='foo6', contents='6')
     with open(fpath_no_longer_used) as f_in:
-      file_md5 = CalculateB64EncodedMd5FromContents(f_in)
+      file_md5 = _CalculateB64EncodedMd5FromContents(f_in)
     mock_api.MockCreateObjectWithMetadata(
         apitools_messages.Object(bucket=bucket_name,
                                  name='foo6', md5Hash=file_md5), contents='6')
@@ -219,7 +221,7 @@ class TestCpFuncs(GsUtilUnitTestCase):
     fpath_uploaded_correctly_url = StorageUrlFromString(
         str(fpath_uploaded_correctly))
     with open(fpath_uploaded_correctly) as f_in:
-      fpath_uploaded_correctly_md5 = CalculateB64EncodedMd5FromContents(f_in)
+      fpath_uploaded_correctly_md5 = _CalculateB64EncodedMd5FromContents(f_in)
     object_uploaded_correctly = mock_api.MockCreateObjectWithMetadata(
         apitools_messages.Object(bucket=bucket_name,
                                  name=fpath_uploaded_correctly,
@@ -254,7 +256,7 @@ class TestCpFuncs(GsUtilUnitTestCase):
     fpath_wrong_contents = self.CreateTempFile(file_name='foo4', contents='4')
     fpath_wrong_contents_url = StorageUrlFromString(str(fpath_wrong_contents))
     with open(self.CreateTempFile(contents='_')) as f_in:
-      fpath_wrong_contents_md5 = CalculateB64EncodedMd5FromContents(f_in)
+      fpath_wrong_contents_md5 = _CalculateB64EncodedMd5FromContents(f_in)
     object_wrong_contents = mock_api.MockCreateObjectWithMetadata(
         apitools_messages.Object(bucket=bucket_name,
                                  name=fpath_wrong_contents,
