+++ gslib/tests/test_hashing_helper.py
@@ -21,12 +21,11 @@ import os
 import pkgutil
 
 from gslib.exception import CommandException
-from gslib.hashing_helper import CalculateMd5FromContents
-from gslib.hashing_helper import HashingFileUploadWrapper
 from gslib.storage_url import StorageUrlFromString
 import gslib.tests.testcase as testcase
 from gslib.utils.constants import TRANSFER_BUFFER_SIZE
-
+from gslib.utils.hashing_helper import CalculateMd5FromContents
+from gslib.utils.hashing_helper import HashingFileUploadWrapper
 
 _TEST_FILE = 'test.txt'
