+++ gslib/tests/test_resumable_streaming.py
@@ -21,13 +21,12 @@ import os
 import pkgutil
 
 from gslib.exception import CommandException
-from gslib.hashing_helper import CalculateHashesFromContents
-from gslib.hashing_helper import CalculateMd5FromContents
 from gslib.resumable_streaming_upload import ResumableStreamingJsonUploadWrapper
 import gslib.tests.testcase as testcase
 from gslib.utils.boto_util import GetJsonResumableChunkSize
 from gslib.utils.constants import TRANSFER_BUFFER_SIZE
-
+from gslib.utils.hashing_helper import CalculateHashesFromContents
+from gslib.utils.hashing_helper import CalculateMd5FromContents
 
 _TEST_FILE = 'test.txt'
