+++ gslib/tests/test_rsync.py
@@ -17,7 +17,6 @@
 import os
 
 import crcmod
-from gslib.hashing_helper import SLOW_CRCMOD_RSYNC_WARNING
 from gslib.project_id import PopulateProjectId
 import gslib.tests.testcase as testcase
 from gslib.tests.testcase.integration_testcase import SkipForGS
@@ -35,6 +34,7 @@ from gslib.tests.util import SetBotoConfigForTest
 from gslib.tests.util import TailSet
 from gslib.tests.util import unittest
 from gslib.utils.boto_util import UsingCrcmodExtension
+from gslib.utils.hashing_helper import SLOW_CRCMOD_RSYNC_WARNING
 from gslib.utils.posix_util import ConvertDatetimeToPOSIX
 from gslib.utils.posix_util import GID_ATTR
 from gslib.utils.posix_util import MODE_ATTR
