+++ gslib/utils/hashing_helper.py
